requirements/requirements_dev.txt
django-debug-toolbar==1.5
django-rest-swagger
pprofile
ipython==5.2.1
unittest2
pep8
flake8
pyflakes
pytest==3.6.0
pytest-cov
pytest-django
pytest-pythonpath
pytest-mock
pytest-timeout
pytest-xdist
logutils
uwsgitop
jupyter
matplotlib
backports.tempfile  # support in unit tests for py32+ tempfile.TemporaryDirectory
mockldap
sdb
